+++ common/db/src/main/java/is/codion/common/db/database/AbstractDatabase.java
@@ -200,6 +200,11 @@ public abstract class AbstractDatabase implements Database {
 	}
 
 	@Override
+	public boolean subqueryRequiresAlias() {
+		return false;
+	}
+
+	@Override
 	public int maximumNumberOfParameters() {
 		return Integer.MAX_VALUE;
 	}